+++ CHANGELOG.md
@@ -19,6 +19,8 @@ minor release, the component will be purged, so be prepared (see `Updating` sect
 - `peapod` command for `neofs-lens` (#2507)
 -  New CLI exit code for awaiting timeout (#2380)
 - Validation of excessive positional arguments to `neofs-cli` commands (#1941)
+- `--lifetime` flag to `bearer create` and `object put` CLI commands  (#1574) 
+- `--expired-at` flag to `session create` and `storagegroup put` CLI commands (#1574)
 
 ### Fixed
 - `neo-go` RPC connection loss handling (#1337)